+++ Options/CHANGELOG.lua
@@ -5,6 +5,7 @@ TMW.CHANGELOG_LASTVER="10.0.0"
 TMW.CHANGELOG = [==[
 ## v11.0.4
 * Fix: #2186 Activation overlays in Retail
+* The cooldown charge sweep is now skinned with Masque.
 
 ## v11.0.3
 * Fix activation overlays in Classic