+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v11.0.5
+* Most cooldown icons with Range Check enabled now utilize more performant event-driven updates in WoW 11.0.
+
 ## v11.0.4
 * Fix: #2186 Activation overlays in Retail
 * The cooldown charge sweep is now skinned with Masque.